+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- The warning shown when using multiple forms with the same id, is now working on `superForm` instantiation, not just when new form data is received.
+
 ## [1.3.0] - 2023-07-14
 
 ### Fixed